updated movies showing now and later
