+++ index.html
@@ -18,6 +18,6 @@
     </div>
     
    
-    <script src="./src/js/movies2.js" type="module"></script>
+    <script src="./src/js/movies.js" type="module"></script>
 </body>
 </html>
