+++ movies.css
@@ -1,3 +1,12 @@
+
+.showing-now, .showing-later{
+    margin: 0 auto;
+    max-width: 1000px;
+    background-color: black;
+    color: white;
+
+}
+
 .movies-now, .movies-later{
    display: grid;
    grid-template-columns: auto auto auto;
@@ -12,3 +21,4 @@ img{
     width: 100%;
     object-fit: cover;
 }
+
