+++ src/main.scss
@@ -1,2 +1,1 @@
 
-@import "movies.scss";
