--- src/movies.scss
@@ -1,14 +1,0 @@
-.movies-now, .movies-later{
-    display: grid;
-    grid-template-columns: auto auto auto;
- }
- 
- .img-div{
-     max-width: 200px;
- }
- 
- img{
-     height: 300px;
-     width: 100%;
-     object-fit: cover;
- }
